builtin/log: die if -k and --suject-extra-prefix are used together
When calling format-patch, -k keeps the commit title as the patch subject
and can't be used with other options that depend on the existence of a
prefix. This way, -k shouldn't also be used with --subject-extra-prefix,
as it introduces an extra prefix.

If -k and --subject-extra-prefix are used together, die.

Signed-off-by: Lucas Seiki Oshiro <lucasseikioshiro@gmail.com>