revision: add subject_extra_field to struct rev_info
Currently, the `struct rev_info` is responsible for holding information
about the prefix that is added at the beginning of the subject when
`format-patch` is called. This way, it makes sense that it also holds
the information of the extra prefix, if it is present.

Add a new field called `subject_extra_field` to `struct rev_info` and
use NULL as its default value.

Signed-off-by: Lucas Seiki Oshiro <lucasseikioshiro@gmail.com>