log-tree: add subject prefix in output email subject
When an extra subject prefix is provided by the user when calling
`format-patch --subject-extra-prefix`, insert it between brackets at the
beginning of the the patch subject. Don't change the behavior when this
flag is not used.

For example, this `format-patch` call:

`git format-patch -1 --subject-extra-prefix EXTRA -n -v2`

will create a patch with the subject beginning with:

`Subject: [EXTRA][PATCH v2 1/1]`

Signed-off-by: Lucas Seiki Oshiro <lucasseikioshiro@gmail.com>