repo-info: add struct repo_info
Introduce the internal struct repo_info. This struct is responsible for
holding the data that will be outputted by `git repo-info` following the
same structure that is outputted by that command.

Add as the first value in this struct `object_format`, currently
retrieved by running `git rev-parse --show-object-format` (sha1 or
sha256), and print its value to stdout.

Add a test asserting that `git repo-info` outputs the object format
correctly.

Signed-off-by: Lucas Seiki Oshiro <lucasseikioshiro@gmail.com>