repo-info: add the --format flag
Currently, git-rev-parse with the 'options for files' flags returns only
the requested values (e.g. `git rev-parse --show-toplevel
--show-ref-format` returns the top-level repository and the reference
storage format), dumping them only separated by line feeds.

However, from the perspective of a programmatic interface, this format
is not a good choice as, for example:

- it doesn't show the keys of the requested fields, meaning that the
  caller depends on the order of the requested fields for parsing them;

- there's no guarantee that each line contains the value of one field,
  for example, the returned values may contain line feeds.

This way, given that git-repo-info aims to produce a machine-readable
output, it needs to follow other formats that are better suited for
those means, for example:

- JSON, which is widely used as a data exchange format;

- null-terminated, composed zero or more <key><LF><value><NUL> entries.

Add the --format flag to the repo-info command, allowing the user to
choose between output formats. Add as options "json" and
"null-terminated", and use "json" as the default format.

Helped-by: Phillip Wood <phillip.wood@dunelm.org.uk>
Helped-by: Junio C Hamano <gitster@pobox.com>
Mentored-by: Karthik Nayak <karthik.188@gmail.com>
Mentored-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Lucas Seiki Oshiro <lucasseikioshiro@gmail.com>