repo: add the flag --format-path
Some paths handled by Git are better presented in their absolute format
and others are better presented relative to the current working
directory.

Add a flag --format-path to git-repo-info, allowing the user to force
the outputted paths to be either in the absolute or in the relative
format, similar to the homonymous flag of git-rev-parsed introduced in
fac60b8925 (rev-parse: add option for absolute or relative path
formatting, 2020-12-13).

Mentored-by: Karthik Nayak <karthik.188@gmail.com>
Mentored-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Lucas Seiki Oshiro <lucasseikioshiro@gmail.com>